requirements.txt
cvxopt==1.3.2
flax==0.10.5
gymnasium==1.1.1
matplotlib==3.10.1
numpy==2.2.4
optax==0.2.4
orbax-checkpoint==0.11.12
pandas==2.2.3
seaborn==0.13.2
stable-baselines3==2.6.0
sb3-contrib==2.6.0
tensorboard==2.19.0
tensorflow-probability==0.25.0
tqdm==4.67.1
wheel==0.45.1
setuptools==75.8.0